Removes jscoverage and coveralls dev dependencies.
jscoverage pulls some dev dependencies that have known CVEs (debug and
uglify-js). There are PRs [1], [2] that fix the CVEs but the project
seems to be unmaintained.

Furthermore, the jscoverage tool is in a broken state at the moment,
most likely because it doesn't understand ES6 (but I could be wrong
about that).

For these two reason I'm booting it from the project and we can
re-evaluate if anything ever changes.

[1]: https://github.com/fishbar/jscoverage/pull/40
[2]: https://github.com/fishbar/jscoverage/pull/41