fix(build) specify extension
Webpack 5 fails with the following error:

~~~~
ERROR in ./node_modules/@jitsi/sdp-interop/lib/index.js 16:0-26
Module not found: Error: Can't resolve './interop' in '/Users/saghul/work/jitsi/lib-jitsi-meet/node_modules/@jitsi/sdp-interop/lib'
Did you mean 'interop.js'?
BREAKING CHANGE: The request './interop' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.
~~~~